+++ app/will/src/main.rs
@@ -8,12 +8,14 @@ extern crate application;
 extern crate application_input;
 extern crate application_robot;
 extern crate application_ui;
+extern crate character_selection_stdio;
 extern crate game_input;
 extern crate game_mode_menu;
 extern crate loading;
 #[macro_use]
 extern crate log;
 extern crate map_loading;
+extern crate map_selection_stdio;
 extern crate object_loading;
 extern crate object_model;
 extern crate stdio_view;
@@ -41,10 +43,12 @@ use application::resource::{
     {self, load_in},
 };
 use application_robot::RobotState;
+use character_selection_stdio::CharacterSelectionStdioBundle;
 use game_input::{GameInputBundle, InputConfig, PlayerActionControl, PlayerAxisControl};
 use game_mode_menu::GameModeMenuState;
 use loading::LoadingState;
 use map_loading::MapLoadingBundle;
+use map_selection_stdio::MapSelectionStdioBundle;
 use object_loading::ObjectLoadingBundle;
 use object_model::config::object::CharacterSequenceId;
 use stdio_view::StdioViewBundle;
@@ -126,6 +130,8 @@ fn run(opt: &Opt) -> Result<(), amethyst::Error> {
             )?.with_bundle(UiBundle::<PlayerAxisControl, PlayerActionControl>::new())?
             .with_bundle(GameInputBundle::new(input_config))?
             .with_bundle(StdioViewBundle::new())?
+            .with_bundle(CharacterSelectionStdioBundle::new())?
+            .with_bundle(MapSelectionStdioBundle::new())?
             .with_bundle(MapLoadingBundle::new())?
             .with_bundle(ObjectLoadingBundle::new())?;
     }