+++ crate/stdio_view/src/stdin_system.rs
@@ -1,16 +1,18 @@
 use std::sync::mpsc::{self, Receiver, TryRecvError};
 use std::thread;
 
-use amethyst::{
-    ecs::prelude::{System, Write},
-    shrev::EventChannel,
-};
+use amethyst::{ecs::prelude::*, shrev::EventChannel};
 use application_input::ApplicationEvent;
+use stdio_spi::VariantAndTokens;
 
 use reader::{self, StdinReader};
+use IoAppEventUtils;
 
 /// Type to fetch the application event channel.
-type EventChannelData<'a> = Write<'a, EventChannel<ApplicationEvent>>;
+type StdinSystemData<'s> = (
+    Write<'s, EventChannel<ApplicationEvent>>,
+    Write<'s, EventChannel<VariantAndTokens>>,
+);
 
 /// Rendering system.
 #[derive(Debug, TypeName)]
@@ -53,15 +55,26 @@ impl Default for StdinSystem {
     } // kcov-ignore
 }
 
-impl<'a> System<'a> for StdinSystem {
-    type SystemData = EventChannelData<'a>;
+impl<'s> System<'s> for StdinSystem {
+    type SystemData = StdinSystemData<'s>;
 
-    fn run(&mut self, mut event_channel: Self::SystemData) {
+    fn run(&mut self, (mut application_event_channel, mut variant_channel): Self::SystemData) {
         match self.rx.try_recv() {
-            Ok(msg) => {
-                debug!("Received message from StdinReader: \"{}\".", msg);
-                if let "exit" = msg.as_str() {
-                    event_channel.single_write(ApplicationEvent::Exit);
+            Ok(input) => {
+                debug!("Input from StdinReader: `{:?}`.", &input);
+
+                if input == "exit" {
+                    application_event_channel.single_write(ApplicationEvent::Exit);
+                    return;
+                }
+
+                match IoAppEventUtils::input_to_variant_and_tokens(input) {
+                    Ok(variant_and_tokens) => {
+                        if let Some(variant_and_tokens) = variant_and_tokens {
+                            variant_channel.single_write(variant_and_tokens);
+                        }
+                    }
+                    Err(e) => error!("Failed to parse input. Error: `{}`.", e),
                 }
             }
             Err(TryRecvError::Empty) => {
@@ -84,84 +97,126 @@ mod test {
         shred::{Resources, SystemData},
         shrev::{EventChannel, ReaderId},
     };
+    use application_event::AppEventVariant;
     use application_input::ApplicationEvent;
 
-    use super::{EventChannelData, StdinSystem};
+    use super::{StdinSystem, StdinSystemData};
+    use stdio_spi::VariantAndTokens;
 
     fn setup() -> (
         StdinSystem,
         Sender<String>,
         Resources,
         ReaderId<ApplicationEvent>,
+        ReaderId<VariantAndTokens>,
     ) {
         let mut res = Resources::new();
         res.insert(EventChannel::<ApplicationEvent>::with_capacity(10));
+        res.insert(EventChannel::<VariantAndTokens>::with_capacity(10));
+
         let (tx, rx) = mpsc::channel();
         let stdin_system = StdinSystem::internal_new(rx, || {});
 
-        let reader_id = {
-            let mut event_channel = EventChannelData::fetch(&res);
-            event_channel.register_reader()
+        let (application_ev_id, variant_and_tokens_id) = {
+            let (mut application_events, mut variant_and_tokens) = StdinSystemData::fetch(&res);
+            (
+                application_events.register_reader(),
+                variant_and_tokens.register_reader(),
+            ) // kcov-ignore
         }; // kcov-ignore
 
-        (stdin_system, tx, res, reader_id)
-    }
-
-    fn expect_event(
-        event_channel: &EventChannel<ApplicationEvent>,
-        mut reader_id: &mut ReaderId<ApplicationEvent>,
-        expected_event: Option<&ApplicationEvent>,
-    ) {
-        let mut event_it = event_channel.read(&mut reader_id);
-        assert_eq!(expected_event, event_it.next());
+        (
+            stdin_system,
+            tx,
+            res,
+            application_ev_id,
+            variant_and_tokens_id,
+        )
     }
 
     #[test]
     fn sends_exit_event_when_input_is_exit() {
-        let (mut stdin_system, tx, res, mut reader_id) = setup();
+        let (mut stdin_system, tx, res, mut application_ev_id, _) = setup();
 
         tx.send("exit".to_string()).unwrap();
         stdin_system.run_now(&res);
 
-        let event_channel = EventChannelData::fetch(&res);
+        let (application_events, _) = StdinSystemData::fetch(&res);
 
         expect_event(
-            &event_channel,
-            &mut reader_id,
+            &application_events,
+            &mut application_ev_id,
             Some(&ApplicationEvent::Exit),
         );
     } // kcov-ignore
 
     #[test]
     fn does_not_send_exit_event_when_input_is_not_exit() {
-        let (mut stdin_system, tx, res, mut reader_id) = setup();
+        let (mut stdin_system, tx, res, mut application_ev_id, _) = setup();
 
         tx.send("abc".to_string()).unwrap();
         stdin_system.run_now(&res);
 
-        let event_channel = EventChannelData::fetch(&res);
-        expect_event(&event_channel, &mut reader_id, None);
+        let (application_events, _) = StdinSystemData::fetch(&res);
+        expect_event(&application_events, &mut application_ev_id, None);
     } // kcov-ignore
 
     #[test]
     fn does_nothing_when_input_is_empty() {
-        let (mut stdin_system, _tx, res, mut reader_id) = setup();
+        let (mut stdin_system, _tx, res, mut application_ev_id, _) = setup();
 
         // we don't call tx.send(..)
         stdin_system.run_now(&res);
 
-        let event_channel = EventChannelData::fetch(&res);
-        expect_event(&event_channel, &mut reader_id, None);
+        let (application_events, _) = StdinSystemData::fetch(&res);
+        expect_event(&application_events, &mut application_ev_id, None);
     } // kcov-ignore
 
     #[test]
-    fn does_not_panic_when_channel_is_disconnected() {
-        let (mut stdin_system, tx, res, mut reader_id) = setup();
+    fn does_not_panic_when_application_channel_is_disconnected() {
+        let (mut stdin_system, tx, res, mut application_ev_id, _) = setup();
 
         drop(tx); // ensure channel is disconnected
         stdin_system.run_now(&res);
 
-        let event_channel = EventChannelData::fetch(&res);
-        expect_event(&event_channel, &mut reader_id, None);
+        let (application_events, _) = StdinSystemData::fetch(&res);
+        expect_event(&application_events, &mut application_ev_id, None);
     } // kcov-ignore
+
+    #[test]
+    fn sends_vat_event_when_input_is_app_event() {
+        let (mut stdin_system, tx, res, _, mut vat_ev_id) = setup();
+
+        tx.send("character_selection confirm".to_string()).unwrap();
+        stdin_system.run_now(&res);
+
+        let (_, vat_events) = StdinSystemData::fetch(&res);
+
+        expect_vat_event(
+            &vat_events,
+            &mut vat_ev_id,
+            Some(&(
+                AppEventVariant::CharacterSelection,
+                vec!["character_selection".to_string(), "confirm".to_string()],
+            )),
+        ); // kcov-ignore
+    }
+
+    fn expect_event(
+        application_events: &EventChannel<ApplicationEvent>,
+        mut application_ev_id: &mut ReaderId<ApplicationEvent>,
+        expected_event: Option<&ApplicationEvent>,
+    ) {
+        let mut event_it = application_events.read(&mut application_ev_id);
+        assert_eq!(expected_event, event_it.next());
+    }
+
+    fn expect_vat_event(
+        vat_events: &EventChannel<VariantAndTokens>,
+        mut vat_ev_id: &mut ReaderId<VariantAndTokens>,
+        expected_event: Option<&VariantAndTokens>,
+    ) {
+        let mut event_it = vat_events.read(&mut vat_ev_id);
+        assert_eq!(expected_event, event_it.next());
+    }
 }